requirements.txt
Flask==3.0.0
ezdxf==1.1.3
PyPDF2==3.0.1
pdfplumber==0.10.3
opencv-python==4.8.1.78
numpy==1.26.2
Pillow==10.1.0

# Windows only (optional, for DWG support)
# Install on Windows: pip install aspose-cad
# aspose-cad==25.12.0

# Linux: DWG support via LibreDWG (system package)
# Install on Linux: compile from https://github.com/LibreDWG/libredwg
